refactor: change file name to camelCase and repath file
